adapted to moved SimuComModel
